+++ lib/src/lib_colors.dart
@@ -2,6 +2,10 @@ import 'package:lib_colors/src/hsl.dart';
 import 'package:lib_colors/src/rgb.dart';
 import 'package:lib_colors/src/hsv.dart';
 
+export 'package:lib_colors/src/hsl.dart';
+export 'package:lib_colors/src/rgb.dart';
+export 'package:lib_colors/src/hsv.dart';
+
 abstract class Color {
   static Color parse(String css) {
     if(css.startsWith('#')) {
@@ -22,7 +26,7 @@ abstract class Color {
   /// Returns HEX representation of the color
   String hex({bool shorten = true, bool withAlpha = true});
 
-  Color get clone;
+  Color clone();
 
   void assignRgb(Rgb rgb);
 
@@ -62,3 +66,13 @@ abstract class Color {
 
   Hsv get toHsv;
 }
+
+Rgb get black => Rgb();
+
+Rgb get white => Rgb(r: 255, g: 255, b: 255);
+
+Rgb get lime => Rgb(g: 255);
+
+Rgb get green => Rgb(g: 128);
+
+Rgb get DarkGreen => Rgb(g: 100);